Ajustes

+++ src/main/java/br/com/aftermidnight/kartwiki/dao/GenericDAO.java
@@ -10,8 +10,8 @@ public interface GenericDAO<T, ID extends Serializable> {
 	public List<T> buscarPorExemplo(T entidade, String... propriedades);
 	public List<T> buscarComConsultaPaginada(String namedQuery, int first, int pageSize, Class<T> clazz);
 	public List<T> buscarComConsultaPaginada(int first, int pageSize);
-	public Integer contarTotalDeRegistros(Class<T> clazz);
-	public Integer contarTotalDeRegistros();
+	public Long contarTotalDeRegistros(Class<T> clazz);
+	public Long contarTotalDeRegistros();
 	
 	public void salvar(T entidade);
 	public void remover(T entidade);
